// src/accounting/transactions/transaction-history/models/transactionHistory.model.ts
export class TransactionHistory {
    amount: number;
    area: string;
    description: string;
    isDeposited: boolean;
    isSplitPayment: boolean;
    locationName: string;
    patientName: string;
    providerUserName: string;
    runningBalance: number;
    tooth: string;
    totalEstimatedInsurance: number;
    type: string;
    utcDateTimeString: string;
    allowedamount : string;
    totalAdjEstimate : string;
};